+++ crates/wdl-engine/tests/inputs.rs
@@ -212,7 +212,9 @@ async fn run_test(test: &Path) -> Result<()> {
                     }
                 }
                 Inputs::Workflow(inputs) => {
-                    let workflow = document.workflow().expect("workflow should be present");
+                    let workflow = document
+                        .local_workflow_by_name(&name)
+                        .expect("workflow should be present");
                     match inputs.validate(document, workflow, None).with_context(|| {
                         format!(
                             "failed to validate the inputs to workflow `{workflow}`",